@@ -6,6 +6,9 @@
     "dev": "next dev --turbopack",
     "dev:tunnel": "concurrently \"next dev --turbopack\" \"npx localtunnel --port 3000 --subdomain myapp\"",
     "tunnel": "npx localtunnel --port 3000 --subdomain myapp",
+    "server": "ts-node app/server.ts",
+    "server:dev": "nodemon --exec ts-node app/server.ts",
+    "dev:all": "concurrently \"npm run dev\" \"npm run server:dev\"",
     "build": "next build",
     "start": "next start",
     "lint": "next lint"
@@ -36,13 +39,16 @@
   "devDependencies": {
     "@eslint/eslintrc": "^3",
     "@tailwindcss/postcss": "^4",
-    "@types/node": "^20",
+    "@types/express": "^5.0.3",
+    "@types/node": "^20.19.14",
     "@types/react": "^19",
     "@types/react-dom": "^19",
+    "concurrently": "^9.2.1",
     "eslint": "^9",
     "eslint-config-next": "15.4.6",
     "tailwindcss": "^4",
+    "ts-node": "^10.9.2",
     "tw-animate-css": "^1.3.6",
-    "typescript": "^5"
+    "typescript": "^5.9.2"
   }
 }
